Added option to deploy kube-router via manifest. This needs to be tested! Also, set kube-router via systemd as default deployment and created default kube-router args.

+++ examples/kube-router/install-kube-router.sh
@@ -3,23 +3,25 @@
 # Starts kube-router on kubernetes workers.
 #######################################
 
-if [[ "${3}" == *"--run-service-proxy=false"* ]]; then
-    echo "Kube-Proxy will be used"
-else
-	echo "Cleaning up any kube-proxy rules before starting kube-router with services handling"
-    sudo kube-proxy --cleanup
-fi
+CNI_INSTALL_TYPE=${1}
+KUBEROUTER_VAGRANT_BIN_DIR=${2}
+ROUTER_ID=${3}
+CNI_ARGS=${4}
 
-
-if systemctl is-active kube-router | grep -q 'inactive'; then
- echo "kube-router not currently running"
-else
-	sudo systemctl stop kube-router
-fi
-
-sudo cp "${1}"kube-router "/usr/bin"
-
-sudo tee /etc/systemd/system/kube-router.service <<EOF
+if [ "${CNI_INSTALL_TYPE}" == "systemd" ]; then
+	if [[ "${CNI_ARGS}" == *"--run-service-proxy=false"* ]]; then
+	    echo "kube-proxy will be used"
+	else
+		echo "Cleaning up any kube-proxy rules before starting kube-router with services handling"
+	    sudo kube-proxy --cleanup
+	fi
+	if systemctl is-active kube-router | grep -q 'inactive'; then
+	 echo "kube-router not currently running"
+	else
+		sudo systemctl stop kube-router
+	fi
+	sudo cp "${KUBEROUTER_VAGRANT_BIN_DIR}"kube-router "/usr/bin"
+	sudo tee /etc/systemd/system/kube-router.service <<EOF
 [Unit]
 Description=Kube-Router
 Documentation=github.com/cloudnativelabs/kube-router
@@ -27,8 +29,8 @@ After=network.target
 
 [Service]
 ExecStart=/usr/bin/kube-router \\
-  --router-id=${2} \\
-  ${3}
+  --router-id=${ROUTER_ID} \\
+  ${CNI_ARGS}
 
 Restart=on-failure
 RestartSec=5
@@ -37,8 +39,12 @@ RestartSec=5
 WantedBy=multi-user.target
 EOF
 
-sudo systemctl daemon-reload
-sudo systemctl enable kube-router
-sudo systemctl restart kube-router
+	sudo systemctl daemon-reload
+	sudo systemctl enable kube-router
+	sudo systemctl restart kube-router
+	sudo systemctl status kube-router --no-pager
 
-sudo systemctl status kube-router --no-pager
+elif [ "${CNI_INSTALL_TYPE}" == "daemonset" ]; then
+	# TODO(arvinder): if --run-service-proxy=true, we likely need to cleanup kube-proxy rules as above
+	kubectl apply -f "${CNI_ARGS}"
+fi
